Notifications for Voicemail Delete Actions (#6263)
As a user, I may want to receive both AMQP and email notifications
when a user deletes a voicemail.

AMQP messages (to the notifications exchange) are sent if voicemails
are deleted via DTMF, delete after notify, and for crossbar actions.

A teletype email template: notification.voicemail_deleted is included
and can be activated by:

The teletype template can be made durable by adding it to
system_config/notify.default.autoload_modules.

Right now, the voicemail delete template is essentially a copy of the
voicemail new template with the Reason added.